:truck: api: Move DomainTransaction to src
