+++ api/src/devcomp/infrastructure/repositories/user-recommended-training-repository.js
@@ -1,5 +1,5 @@
 import { knex } from '../../../../db/knex-database-connection.js';
-import { DomainTransaction } from '../../../../lib/infrastructure/DomainTransaction.js';
+import { DomainTransaction } from '../../../shared/domain/DomainTransaction.js';
 import { UserRecommendedTraining } from '../../domain/read-models/UserRecommendedTraining.js';
 
 const TABLE_NAME = 'user-recommended-trainings';
