+++ api/src/evaluation/application/competence-evaluations/competence-evaluation-controller.js
@@ -1,6 +1,6 @@
 import { evaluationUsecases as usecases } from '../../../evaluation/domain/usecases/index.js';
 import * as competenceEvaluationSerializer from '../../infrastructure/serializers/jsonapi/competence-evaluation-serializer.js';
-import { DomainTransaction } from '../../../../lib/infrastructure/DomainTransaction.js';
+import { DomainTransaction } from '../../../shared/domain/DomainTransaction.js';
 
 const startOrResume = async function (request, h, dependencies = { competenceEvaluationSerializer }) {
   const userId = request.auth.credentials.userId;
