+++ api/src/shared/infrastructure/repositories/assessment-result-repository.js
@@ -1,7 +1,7 @@
 import _ from 'lodash';
 import { knex } from '../../../../db/knex-database-connection.js';
 import { AssessmentResultNotCreatedError, MissingAssessmentId } from '../../domain/errors.js';
-import { DomainTransaction } from '../../../../lib/infrastructure/DomainTransaction.js';
+import { DomainTransaction } from '../../../shared/domain/DomainTransaction.js';
 import { AssessmentResult } from '../../domain/models/AssessmentResult.js';
 import { CompetenceMark } from '../../../../lib/domain/models/CompetenceMark.js';
